Оружие 1.0

+++ bullet.py
@@ -42,11 +42,11 @@ class Bullet(Sprite):
 
     def unit_collided(self, collider, unit):
         if unit.owner in decors or (
-                unit.owner.tag != self.owner.tag and (unit.owner in enemies or unit.owner.tag == "player")):
+                unit.owner.tag != self.owner.tag and (unit.owner in enemies or unit.owner.tag == "player") and unit == unit.owner.colliders["bullet_hit"]):
             self.delete_from_all()
-            print("asdwdasdwd")
         if (unit.owner.tag == "player" or unit.owner.tag == "enemy") and unit.owner.tag != self.owner.tag:
-            unit.owner.hit_from_enemy(self.owner.damage_bullet)
+            if unit == unit.owner.colliders["bullet_hit"]:
+                unit.owner.hit_from_enemy(self.owner.damage_bullet)
 
     def delete_from_all(self):
         self.kill()
